FMI.jl Library Functions

This section lists all functions with their doc strings.

User level

GeoSpatialRelations.intersection — Function

intersection(line::Line, plane::Plane)

Calculating the intersection point between a line and a plane

Arguments

line::Line: instance of a line

plane::Plane: instance of a plane

Returns

intersection_point::Vector: point of intersection

intersection(line::Line, line2::Line)

Calculating the intersection point between two lines

Arguments

line1::Line: instance of a line

line2::Line: instance of a line

Returns

There are two possibilities:

If there is no intersection point, an error message is displayed.

There is one intersection then intersection_point::Vector is returned.

intersection(line::Line, sphere::Sphere)

calculating the intersection point for a line and a sphere

Arguments

line::Line: instance of a line

sphere::Sphere: instance of a sphere

Returns

There are three possibilities:

If there is no intersection point, an error message is displayed.

There is one intersection then intersection_point::Vector is returned.

There are two intersection then intersection_point1::Vector and intersection_point2::Vector are returned.

intersection(circ1::Circle, circ2::Circle)

calculating the intersection point for two cicles

Arguments

circ1::Circle: instance of a circle

circ2::Circle: instance of a circle

Returns

There are three possibilities:

If there is no intersection point, an error message is displayed.

There is one intersection then intersection_point::Vector is returned.

There are two intersection then intersection_point1::Vector and intersection_point2::Vector are returned.

intersection(circ::Circle, sphere::Sphere)

calculating the intersection point for a circle and a sphere

Arguments

circ::Circle: instance of a circle

sphere::Sphere: instance of a sphere

Returns

There are three possibilities:

If there is no intersection point, an error message is displayed.

There is one intersection then intersection_point::Vector is returned.

There are two intersection then intersection_point1::Vector and intersection_point2::Vector are returned.